pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lexalign"
version = "0.1.0"
description = "CLI tool for downloading, fine-tuning, and aligning HuggingFace models"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "huggingface-hub>=0.20.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "trl>=0.9.0,<0.13.0",
    "peft>=0.7.0,<0.15.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.25.0,<2.0.0",
    "datasets>=2.14.0",
    "transformers>=4.36.0,<5.0.0",
    "torch>=2.0.0",
]

[project.scripts]
lexalign-download  = "download:cli"
lexalign-finetune  = "finetune:cli"
lexalign-align     = "align:align"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["lexalign*"]
